mcp-exec/requirements.txt
requests
python-dotenv
langchain
langchain_openai
langchain_community
mcp
langchain-mcp-adapters
langgraph
